backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dotenv==1.0.0
pillow==10.0.0
pdf2image==1.16.3